% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/grasshopper-package.R
\docType{package}
\name{grasshopper}
\alias{grasshopper}
\alias{grasshopper-package}
\title{Grasshopper is an R framework for testing questions about classifiers coming
from learning theory: convergence rate, estimation of empirical risk
minimization potential, etc.}
\description{
Grasshopper is an R framework for testing questions about classifiers coming
from learning theory: convergence rate, estimation of empirical risk
minimization potential, etc.
}

