% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/grasshopper-package.R
\docType{package}
\name{vc.r}
\alias{vc.r}
\alias{vc.r-package}
\title{Analogous to Ruby's VCR gem, this R package
provides the ability to record network responses and grab them from the cache
so that tests are not making repetitive network requests.}
\description{
Analogous to Ruby's VCR gem, this R package
provides the ability to record network responses and grab them from the cache
so that tests are not making repetitive network requests.
}

