% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/package.vc.r.R
\docType{package}
\name{vcr}
\alias{vcr}
\alias{vcr-package}
\title{Analogous to Ruby's VCR gem, this R package
provides the ability to record network responses and grab them from the cache
so that tests are not making repetitive network requests.}
\description{
Analogous to Ruby's VCR gem, this R package
provides the ability to record network responses and grab them from the cache
so that tests are not making repetitive network requests.
}

